requirements.txt
setuptools>=5.5
requests>=2.4
requests_futures>=0.9.5
django-picklefield>=0.3.1
django-bootstrap3>=5.4
lxml>=3.4

